pyproject.toml
[tool.poetry]
name = "mufg-competition"
version = "0.1.0"
description = ""
authors = ["Kusukawa Naoya <kusukawa93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
pandas = "^2.2.2"
scikit-learn = "^1.5.1"
optuna = "^3.6.1"
xgboost = "^2.1.1"
lightgbm = "^4.5.0"
matplotlib = "^3.9.1.post1"
hydra-core = "^1.3.2"
numpy = "1.26.0"
catboost = "^1.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
